Feature addition: Verilog code in Python strings along with verilog files.
The current API is such that it takes in a (list of) file(s) as input. However,
when embedding Pyverilog in aother python libraries/code, it would be helpful
to pass python sitrings directly to the API rather than files.

This feature works along with the current implementation in such a way that
the user can either continue passing files as inputs as in the existing API
or pass verilog code in python strings to the API.